helper for computing library sequences for seqn
